修改登录逻辑

+++ jpower-module-common/common-auth/src/main/java/com/wlcb/jpower/module/common/auth/ClientDetails.java
@@ -20,8 +20,8 @@ public class ClientDetails implements Serializable {
     @ApiModelProperty("客户端密钥")
     private String clientSecret;
     @ApiModelProperty("token有效时长 单位秒")
-    private Integer accessTokenValidity;
+    private Long accessTokenValidity;
     @ApiModelProperty("刷新token有效时长 单位秒")
-    private Integer refreshTokenValidity;
+    private Long refreshTokenValidity;
 
 }
