+++ jpower-module-common/common-auth/src/main/java/com/wlcb/jpower/module/common/auth/TokenInfo.java
@@ -21,6 +21,6 @@ public class TokenInfo {
     /**
      * 过期秒数
      */
-    private int expire;
+    private Long expire;
 
 }
